Rename SvoPropertyBuilder
