+++ src/Qowaiv.EntityFrameworkCore/SvoPropertyBuilder.cs
@@ -1,0 +1,39 @@
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using System.Linq.Expressions;
+
+namespace Qowaiv.EntityFrameworkCore;
+
+/// <summary>Extensions to bind Propereties as SVO's.</summary>
+public static class SvoPropertyBuilder
+{
+    extension<TEntity>(EntityTypeBuilder<TEntity> builder) where TEntity : class
+    {
+        /// <summary>
+        /// Returns a builder that can be used to configure an SVO property of the entity type.
+        /// If the specified property is not already part of the model, it will be added.
+        /// </summary>
+        /// <param name="propertyExpression">
+        /// A lambda expression representing the property to be configured (
+        /// <c>blog => blog.Url</c>).
+        /// </param>
+        [FluentSyntax]
+        [CLSCompliant(false)]
+        public PropertyBuilder<EmailAddress> SvoProperty(Expression<Func<TEntity, EmailAddress>> propertyExpression) => builder
+            .Property(propertyExpression)
+            .HasConversion(email => email.ToString(), str => EmailAddress.Parse(str))
+            .HasMaxLength(EmailAddress.MaxLength);
+    }
+
+    extension(PropertyBuilder<Percentage> builder)
+    {
+        [FluentSyntax]
+        public PropertyBuilder<Percentage> Svo(bool required = true)
+        {
+            Guard.NotNull(builder)
+                .HasConversion(email => (decimal)email, dec => Percentage.Create(dec))
+                .IsRequired(required);
+
+            return builder;
+        }
+    }
+}
